requirements.txt
Django==3.1.5
oauth2-provider==0.0
djangorestframework==3.12.2
social-auth-app-django==4.0.0
django-rest-framework-social-oauth2==1.1.0
